\name{addSplitNames}
\alias{addSplitNames}
\title{
  Add split names to output object in g.part4 and g.part5
}
\description{
  Internal function to add the recording split names to output the reports 
  for GGIR part 4 and 5.
}
\usage{
   addSplitNames(x)
}
\arguments{
  \item{x}{
    Output data.frame as generated inside \link{g.part4} and \link{g.part5}.
  }
}
\keyword{internal}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}