\name{C}
\alias{C}
\docType{data}
\title{
Example output from g.calibrate
}
\description{
C is example output from \link{g.calibrate}
}
\usage{data(C)}
\format{
  The format is:
 chr "C"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(C)
}
\keyword{datasets}
