\name{data.calibrate}
\alias{data.calibrate}
\docType{data}
\title{
Example output from g.calibrate
}
\description{
data.calibrate is example output from \link{g.calibrate}
}
\usage{data(data.calibrate)}
\format{
  The format is:
 chr "data.calibrate"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(data.calibrate)
}
\keyword{datasets}
