\name{data.getmeta}
\alias{data.getmeta}
\docType{data}
\title{
Example output from g.getmeta
}
\description{
data.getmeta is example output from \link{g.getmeta}
}
\usage{data(data.getmeta)}
\format{
  The format is:
 chr "data.getmeta"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(data.getmeta)
}
\keyword{datasets}
