\name{data.inspectfile}
\alias{data.inspectfile}
\docType{data}
\title{
Example output from g.inspectfile
}
\description{
data.inspectfile is example output from \link{g.inspectfile}
}
\usage{data(data.inspectfile)}
\format{
  The format is:
 chr "data.inspectfile"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(data.inspectfile)
}
\keyword{datasets}
