\name{data.metalong}
\alias{data.metalong}
\docType{data}
\title{
  Metalong object as part of part 1 milestone data
}
\description{
  data.metalong is example of the metalong data.frame stored \link{g.part1}
}
\usage{data(data.metalong)}
\format{
  The format is:
 chr "data.metalong"
}
\source{
  The data was collected on one individual for testing purposes
}
\examples{
  data(data.metalong)
}
\keyword{datasets}
