\name{data.ts}
\alias{data.ts}
\docType{data}
\title{
  Time series data.frame stored by part 5
}
\description{
  data.ts is example of the data.frame stored \link{g.part5}
}
\usage{data(data.ts)}
\format{
  The format is:
 chr "data.ts"
}
\source{
  The data was collected on one individual for testing purposes and matches the data
  in object data.metalong
}
\examples{
  data(data.ts)
}
\keyword{datasets}
