\name{extractID}
\alias{extractID}
\title{
  Extract ID from file header object
}
\description{
  Extract ID from file header object.
}
\usage{
   extractID(hvars, idloc, fname)
}
\arguments{
  \item{hvars}{
    Object extracted with \link{g.inspectfile}
  }
  \item{idloc}{
    See \link{GGIR}
  }
  \item{fname}{
    File (base)name.
  }
}
\keyword{internal}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}