\name{filefoldername}
\alias{filefoldername}
\docType{data}
\title{
Example output from g.part1
}
\description{
filefoldername is example output from \link{g.part1}
}
\usage{data(filefoldername)}
\format{
  The format is:
 chr "filefoldername"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(filefoldername)
}
\keyword{datasets}
