\name{filename_dir}
\alias{filename_dir}
\docType{data}
\title{
Example output from g.part1
}
\description{
filename_dir is example output from \link{g.part1}
}
\usage{data(filename_dir)}
\format{
  The format is:
 chr "filename_dir"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(filename_dir)
}
\keyword{datasets}
