\name{g.applymetrics4sg}
\alias{g.applymetrics4sg}
\title{
  Apply metrics specific for data coming out of separategravity function
}
\description{
  Apply metrics specific for data coming out of separategravity function
}
\usage{
  g.applymetrics4sg(acclocal, gvector, 
                                sf, ws3, sg.metrics2do)
}
\arguments{
  \item{acclocal}{
    ...
  }
  \item{gvector}{
    ...
  }
  \item{sf}{
  ...
  }
  \item{ws3}{
  ...
  }
  \item{sg.metrics2do}{
  ...
  }
  
}
\value{
Dataframe with metric values in columns average per epoch (ws3)
}
\author{
Vincent T van Hees <v.vanhees@accelting.com>
}