\name{g.binread}
\alias{g.binread}
\title{
  Deprecated, internally replaced by function readGenea from the GGIRread package
}
\description{
  Function g.binread has been deprecated and has been moved to R package GGIRread and renamed readGenea
}
\usage{
  g.binread(...)
}
\arguments{
  \item{...}{
    Arguments ignored
  }
}