\name{g.cwaread}
\alias{g.cwaread}
\title{
  Deprecated, internally replaced by function readAxivity from the GGIRread package
}
\description{
  Function g.cwaread has been deprecated and has been moved to R package GGIRread and renamed readAxivity
}
\usage{
g.cwaread(...)
}
\arguments{
  \item{...}{
    Arguments ignored
  }
}