\name{g.part5.definedays}
\alias{g.part5.definedays}
\title{
  Fix missing night in part 4 output
}
\description{
  Not intended for direct use by GGIR users.
  Defines when day windows start and end as part of \link{g.part5}.
}
\usage{
  g.part5.definedays(nightsi, wi, indjump, nightsi_bu, 
                              ws3new, qqq_backup=c(), ts, Nts, timewindowi, Nwindows)
}
\arguments{
  \item{nightsi}{
  }
  \item{wi}{
  }
  \item{indjump}{
  }
  \item{nightsi_bu}{
  }
  \item{ws3new}{
  }
  \item{qqq_backup}{
  }
  \item{ts}{
  }
  \item{Nts}{
  }
  \item{timewindowi}{
  }
  \item{Nwindows}{
  }

}
\value{
  List of qqq and qqq_backup 
}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}