\name{g.part5.savetimeseries}
\alias{g.part5.savetimeseries}
\title{
  Saves oart 5 time series to csv files
}
\description{
  Not intended for direct use by GGIR users.
  Saves oart 5 time series to csv files as part of \link{g.part5}.
}
\usage{
  g.part5.savetimeseries(ts, LEVELS, desiredtz, rawlevels_fname)
}
\arguments{
  \item{ts}{
  }
  \item{LEVELS}{
  }
  \item{desiredtz}{
  }
  \item{rawlevels_fname}{
  }
}
\value{
  Function does not provide output, it only prepare data for saving
  and saves it to a file.
}
\author{
  Vincent T van Hees <vincentvanhees@gmail.com>
}