\name{g.plot}
\alias{g.plot}
\title{
function to generate a plot for quality check purposes
}
\description{
Function takes meta-data as generated by \link{g.getmeta} and \link{g.impute}
to create a visual representation of imputed time periods
}
\usage{
g.plot(IMP, M, I, durplot)
}
\arguments{
  \item{IMP}{
  output from \link{g.impute}
}
  \item{M}{
  output from \link{g.getmeta}
}
  \item{I}{
  output from \link{g.inspectfile}
}
  \item{durplot}{
number of days to plot
}
}
\value{
function only produces a plot, no values
}
\author{
  Vincent T van Hees <vincent.van-hees@newcastle.ac.uk>
}