\name{g.report.part4}
\alias{g.report.part4}
\title{
Generate report from milestone data produced by \link{g.part4}
}
\description{
Creates report from milestone data produced by \link{g.part4}. Not intended
  for direct use by package user
}
\usage{
g.report.part4(datadir=c(),metadatadir=c(),loglocation = c(),f0=c(),
  f1=c(),storefolderstructure=TRUE)
  
}
\arguments{
  \item{datadir}{
  see \link{g.part4}
  }
  \item{metadatadir}{
  see \link{g.part4}
  }
  \item{loglocation}{
  see \link{g.part4}
  }
  \item{f0}{
  see \link{g.part4}
  }
  \item{f1}{
  see \link{g.part4}
  }
  \item{storefolderstructure}{
  see \link{g.part4}
  }

}
\value{
Function does not produce data, but only writes reports
in csv format and a visual report in pdf
}
\author{
  Vincent T van Hees <vincentvanhees@gmail.com>
}
