\name{g.wavread}
\alias{g.wavread}
\title{
  Deprecated, internally replaced by function readAxivity from the GGIRread package
}
\description{
  Function g.wavread has been deprecated and has been moved to R package GGIRread and renamed readWav
}
\usage{
  g.wavread(...)
}
\arguments{
  \item{...}{
    Arguments ignored
  }
}