\name{getSplitNames}
\alias{getSplitNames}
\title{
  get split names from filename
}
\description{
  Internal function to get split names from filename if recording was split.
}
\usage{
   getSplitNames(filename)
}
\arguments{
  \item{filename}{
    Character of filename
  }
}
\keyword{internal}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}