\name{GGIR_numUnpack}
\alias{GGIR_numUnpack}
\docType{package}
\title{
  Function using Rcpp
}
\description{
  Function using Rcpp. Not designed for direct interaction by package user
}
\usage{
  GGIR_numUnpack(pack)	
}
\arguments{
  \item{pack}{
    vector of integer
  }
}
\examples{
  \dontrun{
    GGIR_numUnpack()
  }
}
