\name{I}
\alias{I}
\docType{data}
\title{
Example output from g.inspectfile
}
\description{
I is example output from \link{g.inspectfile}
}
\usage{data(I)}
\format{
  The format is:
 chr "I"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(I)
}
\keyword{datasets}
