\name{matchExternalData}
\alias{matchExternalData}
\title{
  Read CGM records collected in the EXTREME project
}
\description{
   Read CGM records collected in the EXTREME project.
}
\usage{
   matchExternalData(externalDatadir = "C:/mystudy/mydata", metadatadir = "C:/mystudy/output_mydata/", colname = NULL, idloc = NULL, desiredtz = "")
}
\arguments{
  \item{externalDatadir}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
  \item{metadatadir}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
  \item{colname}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
  \item{idloc}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
  \item{desiredtz}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
  \item{overwrite}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
  \item{verbose}{
    Character (default = NULL). Path to file with CGM data (accepts txt and xlsx formats).
  }
}
\author{
  Jairo H Migueles <jairo@jhmigueles.com>
}
