\name{tail_expansion_log}
\alias{tail_expansion_log}
\docType{data}
\title{
Example output from g.part1
}
\description{
tail_expansion_log is example output from \link{g.part1}
}
\usage{data(tail_expansion_log)}
\format{
  The format is:
 chr "tail_expansion_log"
}
\source{
The data was collected on one individual for testing purposes
}
\examples{
data(tail_expansion_log)
}
\keyword{datasets}
