\name{createConfigFile}
\alias{createConfigFile}
\title{
Creates Config File based on variables in g.shell.GGIR environment
}
\description{
Only used inside \link{g.shell.GGIR}. Not intended for direct use by user.
}
\usage{
createConfigFile(config.parameters = c())	
}
\arguments{
  \item{config.parameters}{
List with all arguments used in g.shell.GGIR.
}
}
\author{
Vincent T van Hees <v.vanhees@accelting.com>
}