\name{g.part5.addfirstwake}
\alias{g.part5.addfirstwake}
\title{
  Adds first wake if it is missing in part 4 output.
}
\description{
  Not intended for direct use by GGIR users.
  Adds first wake if it is missing in part 4 output
  as part of \link{g.part5}.
}
\usage{
  g.part5.addfirstwake(ts, summarysleep_tmp2, nightsi, sleeplog,
  ID, Nepochsinhour, Nts, SPTE_end, ws3new)
}
\arguments{
  \item{ts}{
  }
  \item{summarysleep_tmp2}{
  }
  \item{nightsi}{
  }
  \item{sleeplog}{
  }
  \item{ID}{
  }
  \item{Nepochsinhour}{
  }
  \item{Nts}{
  }
  \item{SPTE_end}{
  }
  \item{ws3new}{
  }
}
\value{
  Data.frame ts
}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}