\name{g.part5.addsib}
\alias{g.part5.addsib}
\title{
  Adds the sustained inactivity bout to the ts series.
}
\description{
  Not intended for direct use by GGIR users.
  Adds the sustained inactivity bout to the ts series
  as part of \link{g.part5}.
}
\usage{
  g.part5.addsib(ts,ws3, Nts, S2, desiredtz, j, nightsi)
}
\arguments{
  \item{ts}{
  }
  \item{ws3}{
  }
  \item{Nts}{
  }
  \item{S2}{
  }
  \item{desiredtz}{
  }
  \item{j}{
  }
  \item{nightsi}{
  }
}
\value{
  Data.frame ts
}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}