\name{g.part5.lux_persegment}
\alias{g.part5.lux_persegment}
\title{
  Extract key lux variables per segment of the data.
}
\description{
  Extracts per segment of the day: mean lux, time above 1000 lux, time awake,
  and time LUX imputed. Function not intended
  for direct use by package user.
}
\usage{
  g.part5.lux_persegment(ts, sse, LUX_day_segments, ws3new)
}
\arguments{
  \item{ts}{
  }
  \item{sse}{
  }
  \item{LUX_day_segments}{
  }
  \item{ws3new}{
  }
}
\value{
 List with values (vector) of the derived variables and corresponding names (vector).
}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}