\name{g.part5.onsetwaketiming}
\alias{g.part5.onsetwaketiming}
\title{
  Identify wake and sleepperiod window timing
}
\description{
  Not intended for direct use by GGIR users.
  Labels timing of wakeing up and sleep onset as part of \link{g.part5}.
}
\usage{
  g.part5.onsetwaketiming(qqq, ts, min, sec, hour, timewindowi, skiponset, skipwake)
}
\arguments{
  \item{qqq}{
  }
  \item{ts}{
  }
  \item{min}{
  }
  \item{sec}{
  }
  \item{hour}{
  }
  \item{timewindowi}{
  }
  \item{skiponset}{
  }
  \item{skipwake}{
  }

}
\value{
  A list with objects: wake, onset, wakei, onseti, skiponset, and skipwake.
}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}